+++ sdk/modelsrepository/Azure.Iot.ModelsRepository/src/ModelMetadata.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Azure.Iot.ModelsRepository
+namespace Azure.IoT.ModelsRepository
 {
     /// <summary>
     /// ModelMetadata is designated to store KPIs from model parsing.