+++ sdk/modelsrepository/Azure.Iot.ModelsRepository/src/StandardStrings.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Azure.Iot.ModelsRepository
+namespace Azure.IoT.ModelsRepository
 {
     internal class StandardStrings
     {