+++ sdk/modelsrepository/Azure.Iot.ModelsRepository/tests/ModelsRepositoryTestBase.cs
@@ -6,7 +6,7 @@ using System.IO;
 using System.Reflection;
 using System.Text.Json;
 
-namespace Azure.Iot.ModelsRepository.Tests
+namespace Azure.IoT.ModelsRepository.Tests
 {
     /// <summary>
     /// This class will initialize all the settings and create and instance of the ModelsRepoClient.