+++ sdk/modelsrepository/Azure.Iot.ModelsRepository/CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ### New features
 
-- Initial preview of Azure.Iot.ModelsRepository SDK
+- Initial preview of Azure.IoT.ModelsRepository SDK
 
 ### Breaking changes